+++ data.lua
@@ -1,9 +1,0 @@
-data:extend({ 
-  {
-    type = "flying-text",
-    name = "hp-shield",
-    flags = {"not-on-map", "placeable-off-grid"},
-    time_to_live = 2,
-    speed = 0
-  }
-})