+++ show_hp_and_shield/variants/percentage.lua
@@ -5,8 +5,18 @@ percentage.show_hp = function(character)
   if health < 0.98 then
     local surface = character.surface
     local position = character.position
-    local color = {r = 1 - health, g = health, b = 0, a = 0.5}
-    surface.create_entity{name="hp-shield", color = color, text =(math.ceil(health * 100) .. "%"), position= {position.x - 2, position.y - 2.1}}
+    local color = {r = 1 - health, g = health, b = 0, a = 0.8}
+    rendering.draw_text{
+      text = (math.ceil(health * 100) .. "%"),
+      scale = 0.9,
+      surface = character.surface,
+      target = character,
+      color = color,
+      time_to_live = 2,
+      visible = true,
+      alignment = "center",
+      scale_with_zoom = true
+    }
   end
 end
 
@@ -28,13 +38,21 @@ percentage.show_shield = function(character)
   if shield < 0.02 then
     return
   elseif shield < 0.95 then
-    local surface = character.surface
-    local position = character.position
-    local color = {r = 0, g = shield, b = shield, a = 0.5}  
     local abs = math.abs
-    local shield = abs(shield - 1) -- for purple color
-    local color = {r = abs(shield - 1), g = 0, b = 1 - shield, a = 0.5}
-    surface.create_entity{name = "hp-shield", color = color, text = text, position = {position.x + 1.7, position.y - 2.1}}
+    shield = abs(shield - 1) -- for purple color
+    local color = {r = abs(shield - 1), g = 0, b = 1 - shield, a = 0.7}
+    rendering.draw_text{
+      text = text,
+      scale = 0.9,
+      surface = character.surface,
+      target = character,
+      target_offset = {0, 0.3},
+      color = color,
+      time_to_live = 2,
+      visible = true,
+      alignment = "center",
+      scale_with_zoom = true
+    }
   end
 end
 
@@ -56,13 +74,19 @@ percentage.show_shield_for_vehicles = function(vehicle)
   if shield < 0.02 then
     return
   elseif shield < 0.95 then
-    local surface = entity.surface
-    local position = entity.position
-    local color = {r = 0, g = shield, b = shield, a = 0.5}
     local abs = math.abs
-    local shield = abs(shield - 1) -- for purple color
-    local color = {r = abs(shield - 1), g = 0, b = 1 - shield, a = 0.5}
-    global.vehicles_shield[tostring(entity.unit_number)].text = surface.create_entity{name = "hp-shield", color = color, text = text, position = {position.x + 1.7, position.y - 2.1}}
+    shield = abs(shield - 1) -- for purple color
+    local color = {r = abs(shield - 1), g = 0, b = 1 - shield, a = 0.7}
+    rendering.draw_text{
+      text = text,
+      surface = vehicle.surface,
+      target = vehicle,
+      color = color,
+      time_to_live = 2,
+      visible = true,
+      alignment = "center",
+      scale_with_zoom = true
+    }
   end
 end