+++ docs/ru/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2019-03-02
+
+### [v2.3.0][v2.3.0]
+
+* Добавлено: индивидуальная визуализация
+* Изменено: только товарищ по команде может видить ваше здоровоье и видеть щит транспорта вашей фракции
+
+### [v2.2.0][v2.2.0]
+
+* Новая визуализация
+
 ## 2019-03-01
 
 ### [v2.1.0][v2.1.0]
@@ -121,6 +132,8 @@
 
 * Первый релиз для 0.16
 
+[v2.3.0]: https://mods.factorio.com/mod/show-health-and-shield/downloads
+[v2.2.0]: https://mods.factorio.com/mod/show-health-and-shield/downloads
 [v2.1.0]: https://mods.factorio.com/download/show-health-and-shield/5c78e511507d61000dcd440d
 [v2.0.0]: https://mods.factorio.com/download/show-health-and-shield/5c320695199eec000d30270d
 [v1.3.20]: https://mods.factorio.com/download/show-health-and-shield/5ad8821f6e5059000a8ae9e8